+++ src/main/java/br/gov/caixa/domain/model/Simulacao.java
@@ -1,8 +1,10 @@
 package br.gov.caixa.domain.model;
 
+import java.math.BigDecimal;
+
 public record Simulacao(
         Long clienteId,
-        Double valor,
+        BigDecimal valor,
         Integer prazoMeses,
         String tipoProduto
 ) {
